doscount set to integer value

+++ database/migrations/2025_07_01_103900_add_new_fields_to_products_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
                   ->default('everyday_essential')
                   ->after('sub_category_id');
             $table->string('thumbnail_image')->nullable()->after('brand');
-            $table->decimal('discount_percentage', 5, 2)->default(0)->after('price');
+            $table->integer('discount_percentage')->default(0)->after('price');
             $table->decimal('discounted_price', 10, 2)->nullable()->after('discount_percentage');
             $table->boolean('has_variants')->default(false)->after('status');
         });
